+++ internal/invapi/portal.go
@@ -5,7 +5,7 @@ const (
 	DefaultBaseURL        = "https://invapi.hostkey.com/"
 	PortalDomain          = "hostkey.com"
 	SiblingPortalDomain   = "hostkey.ru"
-	SiblingProviderSource = "hostkey-cloud/hostkey-ru"
+	SiblingProviderSource = "hostkey-cloud-ru/hostkey-ru"
 	SiblingAPIHostHint    = "invapi.hostkey.ru"
 	DocsAccountKeysURL    = "https://hostkey.com/documentation/account/api_key_account/"
 	ProviderBinaryName    = "terraform-provider-hostkey-com"